+++ src/Yasumi/Provider/Spain.php
@@ -83,7 +83,10 @@ class Spain extends AbstractProvider
         if ($this->year >= 1981) {
             $this->addHoliday(new Holiday(
                 'nationalDay',
-                ['es' => 'Fiesta Nacional de España'],
+                [
+                    'ca' => 'Festa Nacional d\'Espanya',
+                    'es' => 'Fiesta Nacional de España',
+                ],
                 new DateTime("$this->year-10-12", new DateTimeZone($this->timezone)),
                 $this->locale
             ));
@@ -109,7 +112,10 @@ class Spain extends AbstractProvider
         if ($this->year >= 1978) {
             $this->addHoliday(new Holiday(
                 'constitutionDay',
-                ['es' => 'Día de la Constitución'],
+                [
+                    'ca' => 'Dia de la Constitució',
+                    'es' => 'Día de la Constitución',
+                ],
                 new DateTime("$this->year-12-6", new DateTimeZone($this->timezone)),
                 $this->locale
             ));