+++ src/Yasumi/Provider/Spain/BalearicIslands.php
@@ -80,7 +80,10 @@ class BalearicIslands extends Spain
         if ($this->year >= 1983) {
             $this->addHoliday(new Holiday(
                 'balearicIslandsDay',
-                ['es' => 'Día de les Illes Balears'],
+                [
+                    'ca' => 'Diada de les Illes Balears',
+                    'es' => 'Día de les Illes Balears',
+                ],
                 new DateTime("$this->year-3-1", new DateTimeZone($this->timezone)),
                 $this->locale
             ));