+++ src/Yasumi/data/translations/stJohnsDay.php
@@ -13,6 +13,7 @@
 
 // Translations for St. John's Day
 return [
+    'ca' => 'Sant Joan',
     'da' => 'sankthansaften',
     'el' => 'Σύναξις Προφήτου Προδρόμου και Βαπτιστού Ιωάννου',
     'en' => 'St. John\'s Day',