+++ src/Yasumi/data/translations/stJosephsDay.php
@@ -13,6 +13,7 @@
 
 // Translations for St. Joseph's Day
 return [
+    'ca' => 'Sant Josep',
     'de' => 'Josephstag',
     'en' => 'St. Joseph\'s Day',
     'es' => 'San José',